+++ core/spec/fakes/order.rb
@@ -1,0 +1,7 @@
+require 'fakes/model'
+
+module Spree
+  class Order
+    include FakeModel
+  end
+end
