--- core/spec/support/fake_order.rb
@@ -1,6 +1,0 @@
-require 'support/fake_model'
-
-module Spree
-  class Order < FakeModel
-  end
-end
